requirements.txt
Levenshtein
lxml
beautifulsoup4
requests
python-shout
dynaconf
yt-dlp